LayoutTests/printing/page-rule-selection-expected.txt

This tests page style selection.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Default margin is auto.
PASS testRunner.pageProperty('margin-left', 0) is "auto"
PASS testRunner.pageProperty('margin-left', 1) is "auto"
PASS testRunner.pageProperty('margin-left', 2) is "auto"
PASS testRunner.pageProperty('margin-left', 3) is "auto"
PASS testRunner.pageProperty('margin-left', 4) is "auto"
Set margin to 200px via user style sheet.
PASS testRunner.pageProperty('margin-left', 0) is "200"
PASS testRunner.pageProperty('margin-left', 1) is "200"
PASS testRunner.pageProperty('margin-left', 2) is "200"
PASS testRunner.pageProperty('margin-left', 3) is "200"
PASS testRunner.pageProperty('margin-left', 4) is "200"
Set margin to 100px.
PASS testRunner.pageProperty('margin-left', 0) is "100"
PASS testRunner.pageProperty('margin-left', 1) is "100"
PASS testRunner.pageProperty('margin-left', 2) is "100"
PASS testRunner.pageProperty('margin-left', 3) is "100"
PASS testRunner.pageProperty('margin-left', 4) is "100"
Set margin for right pages to 50px. The first page is a right page.
PASS testRunner.pageProperty('margin-left', 0) is "50"
PASS testRunner.pageProperty('margin-left', 1) is "100"
PASS testRunner.pageProperty('margin-left', 2) is "50"
PASS testRunner.pageProperty('margin-left', 3) is "100"
PASS testRunner.pageProperty('margin-left', 4) is "50"
Set margin for left pages to 75px.
PASS testRunner.pageProperty('margin-left', 0) is "50"
PASS testRunner.pageProperty('margin-left', 1) is "75"
PASS testRunner.pageProperty('margin-left', 2) is "50"
PASS testRunner.pageProperty('margin-left', 3) is "75"
PASS testRunner.pageProperty('margin-left', 4) is "50"
Set margin for the first page to 125px.
PASS testRunner.pageProperty('margin-left', 0) is "125"
PASS testRunner.pageProperty('margin-left', 1) is "75"
PASS testRunner.pageProperty('margin-left', 2) is "50"
PASS testRunner.pageProperty('margin-left', 3) is "75"
PASS testRunner.pageProperty('margin-left', 4) is "50"
Change the document direction to RTL. This makes the first page a left page.
PASS testRunner.pageProperty('margin-left', 0) is "125"
PASS testRunner.pageProperty('margin-left', 1) is "50"
PASS testRunner.pageProperty('margin-left', 2) is "75"
PASS testRunner.pageProperty('margin-left', 3) is "50"
PASS testRunner.pageProperty('margin-left', 4) is "75"
Check font update and line height override.
PASS testRunner.pageProperty('line-height', 0) is "180"
PASS testRunner.pageProperty('font-family', 0) is "Arial"
PASS testRunner.pageProperty('font-size', 0) is "30"
Check @media rules.
PASS testRunner.pageProperty('size', 0) is "100 200"
PASS testRunner.pageProperty('size', 0) is "150 250"
PASS testRunner.pageProperty('size', 0) is "150 250"
PASS successfullyParsed is true

TEST COMPLETE
